Moved ColumnTransformer to data-model as well
